creates IconButton component; moves Button.svelte into new directory
