+++ src/lib/components/modal/ModalAction.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Button from "../Button.svelte";
+  import { Button } from "$lib/components/button";
 
   export let primary = false;
 </script>
